+++ Makefile
@@ -6,7 +6,7 @@
 #    By: gsmereka <gsmereka@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/01/20 18:26:17 by gde-mora          #+#    #+#              #
-#    Updated: 2023/01/30 12:10:46 by gsmereka         ###   ########.fr        #
+#    Updated: 2023/02/03 12:17:12 by gsmereka         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -14,6 +14,7 @@ NAME 	=	minishell
 
 SRC 	=	src/main.c \
 			src/repl/init_repl.c \
+			src/repl/init_repl_signals.c \
 
 OBJ 	=	$(addprefix $(OBJ_DIR)/, $(SRC:.c=.o))
