+++ headers/functions.h
@@ -6,22 +6,21 @@
 /*   By: gsmereka <gsmereka@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/23 15:23:28 by gde-mora          #+#    #+#             */
-/*   Updated: 2023/01/30 13:11:05 by gsmereka         ###   ########.fr       */
+/*   Updated: 2023/02/03 12:16:32 by gsmereka         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FUNCTIONS_H
 # define FUNCTIONS_H
 # include "structs.h"
-# include <signal.h>
 
 // REPL
+# include <signal.h>
 # include <stdio.h>
 # include <readline/readline.h>
 # include <readline/history.h>
 
 void	init_repl(t_data *data);
-
-// Signals
+void	init_repl_signals(t_data *data);
 
 #endif
